+++ kibot/gs.py
@@ -216,6 +216,7 @@ class GS(object):
     global_use_dir_for_preflights = None
     global_use_os_env_for_expand = None
     global_variant = None
+    global_vrml_3d_model_workaround = None
     # Only for v7+
     global_allow_blind_buried_vias = None
     global_allow_microvias = None